requirements.txt
pyupbit>=0.2.33
python-dotenv>=1.0.0
requests>=2.31.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
aiosqlite>=0.19.0
